Removed the exit alias in runshell function.
It was the culprit of the git error message apearing in the do) verb and was not doing what it was supposed to do anyway.
The alias was checking if git status was clean when entering the shell, not when leaving.
Now the status is not checked and it is possible to exit the shell with uncommitted work.