Verbs like done), kill), move), new), redo), rename) and undone) enable and disable projects accordingly.
It's natural that you might want to work on projects you create as a new project or redo an old one or if you decide it was undone.
Likewise, done and killed projects are not the ones you want in your working directory anymore.
Links to moved or renamed projects are updated in conformity to theyre new and previous state.